Move JsonRowsFormatter under TurboTests
